#415: ask the compiler where it put the fields, for both targets
The half of this audit that no warning reaches is the structures that cross a
boundary.  Both sides compile, the disk or the wire or the shared page does
not, and the failure is silent.  ELF and ext2 were found by following a
format warning into the header next door, which is luck rather than method.

scripts/layout-diff.py is the method.  It asks clang for its record layouts
at -m32 and -m64 and diffs them, field by field.  clang is the instrument and
not the compiler here -- the kernel is still built with gcc, and nothing this
prints goes into it -- but gcc has no way to answer the question and clang
does.

Two mistakes made while writing it, both of which produced confident wrong
answers, and both now written into the tool's own comments:

  The include path has to differ per target.  Compiling at -m64 against the
  i386 machine headers leaves vm_offset_t thirty-two bits wide, so mach/
  boot_info.h came back SAME when in fact every one of its structures
  doubles.  A measurement that uses the wrong headers is not a weaker
  measurement, it is a different one.

  clang prints a layout only for records whose layout is demanded, so the
  probe has to take sizeof rather than declare a variable.  Before that it
  reported "0 differ" for headers it had not looked at.

Controls, since "0 differ" has to be earned: the two headers already fixed
come back SAME, and the pre-fix ext2 header comes back with all six records
differing -- superblock 1024 to 2040, inode 128 to 256, group descriptor 32
to 56, which is the damage the previous commit described and now measures.

What the sweep found, classified rather than just listed:

  by design      mach/message.h -- descriptors 12 to 16, body 4 to 8.  That
                 is #413's ABI, not a defect.

  safe as built  the MIG *_info structures.  They differ, but each travels
                 with a count computed from its own size, and #413 already
                 ruled out a 32-bit userland on a 64-bit kernel.

  not a defect   cap_manifest_entry, same size and same offsets, differing
                 only in alignment because of a uint64_t.  The explicit _pad
                 says someone had already thought about it.

  dead           mach/boot_info.h.  Its four structures all differ, and
                 move_bootstrap() is called only from hp_pa/, which is not a
                 target and is not compiled.  Left alone, and said so.

  fixed here     host_cpu_load_info and processor_cpu_load_info declared
                 their tick counters `unsigned long'.  These travel as an
                 array of natural_t whose length is sizeof/sizeof(natural_t),
                 so on x86-64 each counter would have taken two elements
                 where the interface describes one.  Nothing breaks while
                 both sides are built together, but the structure stops being
                 the one the interface describes, and cpustat reads it from
                 userland.  Now unsigned int, and both targets agree.

Both trees build.  Remaining boundaries for this sweep: the bundle format,
GPT, and /proc, which is text and lower risk.

+++ uros/uapi/mach/host_info.h
@@ -214,8 +214,17 @@ typedef struct host_load_info	*host_load_info_t;
 #define	HOST_VM_INFO_COUNT \
 		(sizeof(vm_statistics_data_t)/sizeof(integer_t))
 
+/*
+ * #415: unsigned int, which is what Mach's ABI says and what the transport
+ * assumes.  These travel as an array of natural_t whose length is
+ * sizeof(this)/sizeof(natural_t), so with `unsigned long' each counter
+ * occupied two elements on x86-64 and one here -- the count adjusts, so
+ * nothing breaks while both sides are built for the same target, but the
+ * structure stops being the one the interface describes and every reader
+ * that walks the array as integers sees halves of counters.
+ */
 struct host_cpu_load_info {		/* number of ticks while running... */
-	unsigned long	cpu_ticks[CPU_STATE_MAX]; /* ... in the given mode */
+	unsigned int	cpu_ticks[CPU_STATE_MAX]; /* ... in the given mode */
 };
 typedef struct host_cpu_load_info	host_cpu_load_info_data_t;
 typedef struct host_cpu_load_info	*host_cpu_load_info_t;
